requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.5
flask-migrate==2.3.0
requests==2.20.1
flask-cors
